[Refactor] 패키지 경로를 http 로 수정
