+++ src/main/java/codesquad/http/message/InvalidResponseFormatException.java
@@ -1,0 +1,7 @@
+package codesquad.http.message;
+
+public class InvalidResponseFormatException extends RuntimeException{
+    public InvalidResponseFormatException() {
+        super("Invalid resopnse format");
+    }
+}
