+++ src/main/java/codesquad/http/message/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package codesquad.message.request;
+package codesquad.http.message.request;
 
 public enum HttpMethod {
     GET("GET"),POST("POST"),PUT("PUT"),DELETE("DELETE"),PATCH("PATCH"),HEAD("HEAD"),OPTIONS("OPTIONS"),TRACE("TRACE"),CONNECT("CONNECT");
