--- src/main/java/codesquad/message/InvalidResponseFormatException.java
@@ -1,7 +1,0 @@
-package codesquad.message;
-
-public class InvalidResponseFormatException extends Exception{
-    public InvalidResponseFormatException() {
-        super("Invalid resopnse format");
-    }
-}
