+++ src/test/java/codesquad/message/request/HttpMethodTest.java
@@ -1,5 +1,6 @@
 package codesquad.message.request;
 
+import codesquad.http.message.request.HttpMethod;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
